server1/requirements.txt
Flask
geopy
opencv-python
Pillow
pandas
gunicorn
werkzeug
